fix: routing issue nextjs
