+++ apps/app/dashboards/democracy/election-explorer/candidates.tsx
@@ -165,7 +165,7 @@ const ElectionCandidatesDashboard: FunctionComponent<ElectionCandidatesProps> = 
             {
               name: t("elections"),
               icon: <SPRIconSolid className="-mb-1" />,
-              url: routes.ELECTION_EXPLORER,
+              url: routes.ELECTION_EXPLORER.concat("/elections"),
             },
             {
               name: t("candidates"),
