+++ apps/app/dashboards/democracy/election-explorer/seats.tsx
@@ -142,7 +142,7 @@ const ElectionSeatsDashboard: FunctionComponent<ElectionSeatsProps> = ({
             {
               name: t("elections"),
               icon: <SPRIconSolid className="-mb-1" />,
-              url: routes.ELECTION_EXPLORER,
+              url: routes.ELECTION_EXPLORER.concat("/elections"),
             },
             {
               name: t("candidates"),
