+++ apps/app/pages/dashboard/election-explorer/elections/[...election].tsx
